+++ sdk/communication/Azure.Communication.PhoneNumbers/CHANGELOG.md
@@ -17,15 +17,13 @@
 - PurchasePhoneNumbersOperation and ReleasePhoneNumberOperation extend Operation instead of Operation<Response>.
 - Removed PhoneNumberOperationStatus and PhoneNumberOperationType.
 - Renamed ISOCurrencySymbol property to IsoCurrencySymbol in PhoneNumberCost.
-- Renamed threeLetterISOCountryName parameter to threeLetterIsoCountryName in PhoneNumbersClient.StartSearchAvailablePhoneNumbers and PhoneNumbersClient.StartSearchAvailablePhoneNumbersAsync.
+- Renamed threeLetterISOCountryName parameter to twoLetterIsoCountryName in PhoneNumbersClient.StartSearchAvailablePhoneNumbers and PhoneNumbersClient.StartSearchAvailablePhoneNumbersAsync.
 
 ## 1.0.0-beta.5 (2021-03-09)
 
 ### Added
 - Added PhoneNumbersClient (originally was part of the Azure.Communication.Administration package).
 - Added support for Azure Active Directory Authentication.
-
-
 
 ### Breaking Changes
 - PhoneNumberAdministrationClient has been replaced with PhoneNumbersClient, which has the same functionality but different APIs. To learn more about how PhoneNumbersClient works, refer to the [README.md][read_me]